@@ -15,27 +15,25 @@ keywords =
     sdc, privacy, anonymization, data, automation, perturbation, generalization, suppression, optimization
 
 [options]
-python_requires = >= 3.8
+python_requires = >= 3.10
 packages = find:
 package_dir =
        = src
 install_requires =
-       boruta>=0.3
-       bpemb>=0.3.3
-       matplotlib>=3.4.2
-       numpy>=1.19.3
-       pandas>=1.1.3
+       bpemb~=0.3
+       matplotlib~=3.5
+       numpy~=1.26
+       pandas~=1.4
        plotly>=4.11.0
-       scikit-learn>=0.23
-       scipy>=1.5.4
-       statsmodels>=0.12
-       SALib>=1.4.5
-       tensorflow>=2.4.0
-       tpot[dask]>=0.11
+       SALib~=1.4
+       scikit-learn~=1.1
+       scipy~=1.8
+       statsmodels~=0.13
+       tensorflow~=2.9
+       tpot[dask]~=0.11
 tests_require = 
-       pytest>=3.8
-       hypothesis>=4.53.2
-       beartype>=0.5.1
+       beartype>=0.5.1  
+       pytest~=7.4
 
 [options.packages.find]
 where = src
@@ -47,7 +45,7 @@ python_files=test_*.py
 testpaths=tests
 
 [tox:tox]
-envlist = py38, py39, coverage, bandit, owasp-depcheck
+envlist = py3{10,11}, coverage, bandit, owasp-depcheck
 toxworkdir = build/tox
 
 [testenv]
@@ -58,7 +56,7 @@ commands = pytest tests --junitxml={toxworkdir}/xunit-tests-{envname}.xml -o jun
 
 [testenv:coverage]
 usedevelop = true
-basepython = python3.8
+basepython = python3.10
 deps = {[testenv]deps}
        coverage
        pytest-cov
@@ -66,16 +64,16 @@ commands = pytest --cov-report xml:{toxworkdir}/xunit-coverage.xml --cov-config=
 
 [testenv:localcoverage]
 usedevelop = true
-basepython = python3.8
+basepython = python3.10
 deps = {[testenv]deps}
        coverage
        pytest-cov
 commands = pytest --cov-report term-missing --cov-config=setup.cfg --cov=pymasq tests
 
 [testenv:bandit]
-basepython = python3.8
+basepython = python3.10
 deps = bandit
 commands = bandit -f json -o {toxworkdir}/security-bandit.json -r {envsitepackagesdir}/pymasq
 
 [testenv:owasp-depcheck]
-basepython = python3.8
+basepython = python3.10
