@@ -77,11 +77,12 @@ def load_loan():
     """
     return load_data("loan.csv")
 
+
 def load_bank_attrition_rates():
     """Load and return the Bank Attrition Rates dataset.
 
-    A manager at the bank is disturbed with more and more customers leaving their credit card services. 
-    They would really appreciate if one could predict for them who is gonna get churned so 
+    A manager at the bank is disturbed with more and more customers leaving their credit card services.
+    They would really appreciate if one could predict for them who is gonna get churned so
     they can proactively go to the customer to provide them better services and turn customers' decisions in the opposite direction.
 
     ==============   ==============
