@@ -1,34 +1,34 @@
-
 """
 Expose public exceptions & warnings
 """
 
+
 class InputError(Exception):
-    """ Exception raised for errors in the input value. """
+    """Exception raised for errors in the input value."""
 
 
 class DataTypeError(Exception):
-    """ Exception raised for errors in the data type. """
-    
-    
+    """Exception raised for errors in the data type."""
+
+
 class SumNotEqualToOneError(ValueError):
-    """ Exception for sum of values not equal to 1. """
-    
+    """Exception for sum of values not equal to 1."""
+
 
 class NotInRangeError(ValueError):
-    """ Exception for values not in specified interval. """
+    """Exception for values not in specified interval."""
 
 
 class LessThanZeroError(ValueError):
-    """ Exceptions for values < 0. """
+    """Exceptions for values < 0."""
 
 
 class LessThanOrEqualToZeroError(ValueError):
-    """ Exceptions for values <= 0. """
+    """Exceptions for values <= 0."""
 
 
 class NoMutationAvailableError(ValueError):
-    """ Exception when all mutations have been discarded and not replaced """
+    """Exception when all mutations have been discarded and not replaced"""
 
 
 __all__ = [
@@ -38,5 +38,5 @@ __all__ = [
     "NotInRangeError",
     "LessThanZeroError",
     "LessThanOrEqualToZeroError",
-    "NoMutationAvailableError"
+    "NoMutationAvailableError",
 ]
