@@ -1,3 +1,4 @@
+import logging
 import math
 from typing import Union, List, Final, Optional
 
@@ -7,10 +8,10 @@ import numpy as np
 import scipy.stats as ss
 
 from pymasq import BEARTYPE
-from pymasq.config import FORMATTING_ON_OUTPUT
+from pymasq.config import DEFAULT_SEED, FORMATTING_ON_OUTPUT
 from pymasq.utils import formatting
-from pymasq.preprocessing import LabelEncoder_pm
-from pymasq.errors import InputError, DataTypeError
+from pymasq.preprocessing import LabelEncoderPM
+from pymasq.errors import InputError
 from pymasq.mitigations.utils import _is_identical
 
 
@@ -25,12 +26,15 @@ __all__ = [
     "MODEL",
 ]
 
+logger = logging.getLogger(__name__)
 
 SPEARMAN: Final = "spearman"
 PEARSON: Final = "pearson"
 KENDALL: Final = "kendall"
 CORRELATIVE: Final = "corr"
 MODEL: Final = "model"
+
+rg = np.random.default_rng(DEFAULT_SEED)
 
 
 @BEARTYPE
@@ -210,11 +214,11 @@ def shuffle(
                 f"The values of `data[{cor_cols}]` are all identical and therefore cannot be used for correlation."
             )
         else:
-            print(
+            logger.info(
                 "WARNING: ignoring columns that are composed entirely of identical values."
             )
 
-    _data = LabelEncoder_pm.encode(df=pd.concat([x, y], axis=1))
+    _data = LabelEncoderPM.encode(df=pd.concat([x, y], axis=1))
 
     resp_cols = y.columns.to_list()
     pred_cols = x.columns.to_list()
@@ -232,7 +236,7 @@ def shuffle(
     ystar1 = predictors.dot(pxs.dot(pssinv).T)
 
     sigma = pxx - pxs.dot(pssinv.dot(psx))
-    e1 = np.random.multivariate_normal(
+    e1 = rg.multivariate_normal(
         mean=[0] * len(resp_cols), cov=sigma, size=_data.shape[0]
     )
     y_star = ystar1 + e1
