@@ -1,11 +1,10 @@
 import pandas as pd
 
-from typing import Union, List, Optional, Dict, Final
+from typing import Union, List, Optional, Final
 
 from pymasq import BEARTYPE
 from pymasq.config import (
     FORMATTING_ON_OUTPUT,
-    FORMATTING_IGNORE_DTYPES,
     VALIDATE_NUMERIC_ON_INPUT,
     VALIDATE_NUMERIC_ON_OUTPUT,
 )
