@@ -1,11 +1,11 @@
 from .preprocess import (
     EmbeddingsEncoder,
-    LabelEncoder_pm,
+    LabelEncoderPM,
 )
 from .entity_embedding import embed_entities
 
 __all__ = [
     "embed_entities",
     "EmbeddingsEncoder",
-    "LabelEncoder_pm",
+    "LabelEncoderPM",
 ]
