@@ -1,17 +1,19 @@
 #!/usr/bin/env python
 # coding: utf-8
 
+import logging
+import hashlib
+
 import pytest
 
-import hashlib
-import numpy as np
-
+from pymasq.config import rg
 from pymasq.datasets import load_census
 from pymasq.mitigations import hashing
 
 
 ALGORITHMS = hashlib.algorithms_guaranteed
 
+logger = logging.getLogger(__name__)
 
 def _my_df():
     df = load_census()
@@ -28,7 +30,7 @@ def my_df():
 @pytest.fixture
 def salts():
     df = _my_df()
-    return np.random.choice(["a", "b", "c"], size=df.shape).tolist()
+    return rg.choice(["a", "b", "c"], size=df.shape).tolist()
 
 
 @pytest.mark.parametrize("hash_func", (ALGORITHMS))
@@ -36,12 +38,12 @@ def test_hashing_all_hashlib_guaranteed_algorithms(my_df, hash_func):
     """
     Test all hashing algorithms that are guaranteed to be supported by hashlib, regardless of OS platform.
     """
-    e = None
+    rdf = None
     try:
         rdf = hashing(my_df, hash_func)
     except Exception as e:
-        print("Raised Exception")
-    assert e is None
+        logger.info(f"Raised Exception: {e}")
+    assert rdf is not None
 
 
 @pytest.mark.parametrize("hash_func", (ALGORITHMS))
