@@ -6,9 +6,8 @@ from pymasq.datasets import load_data, load_loan
 from pymasq.mitigations import (
     s,  # shuffle.py module
     shuffle,
-    MODEL,
 )
-from pymasq.errors import InputError, DataTypeError
+from pymasq.errors import InputError
 
 
 @pytest.fixture
@@ -88,7 +87,7 @@ def test_shuffle_cols_not_numeric(my_df):
 
 
 def test_shuffle_same_mean_different_values(loan_df):
-    """ Test that values are perturbed and retain the same mean while also in different order """
+    """Test that values are perturbed and retain the same mean while also in different order"""
     shuffle_cols = ["ApplicantIncome", "LoanAmount"]
     cor_cols = ["Education", "Loan_Status"]
     shuffled = shuffle(
